fpu.c修正

+++ fpu/fpu.c
@@ -85,7 +85,7 @@ float fadd(float f1,float f2){
 }
 
 float fmul(float f1,float f2){
-  unsigned long fraction = 0;
+  unsigned long long fraction = 0;
   unsigned exp = 0;
   ieee_f f1_b,f2_b,f_b;
   f_b.f = 0;
@@ -100,7 +100,7 @@ float fmul(float f1,float f2){
     return f_b.f;
   }
 
-  fraction = (unsigned long)(0x800000+f1_b.b.fraction) * (unsigned long)(0x800000+f2_b.b.fraction);
+  fraction = (unsigned long long)(0x800000+f1_b.b.fraction) * (unsigned long long)(0x800000+f2_b.b.fraction);
   exp = f1_b.b.exp + f2_b.b.exp + 129;
 
   if (fraction < 0x800000000000) { //仮数部の乗算が2.0未満
@@ -124,8 +124,8 @@ float fmul(float f1,float f2){
 }
 
 float fdiv(float f1,float f2){
-  unsigned long fraction = 0;
-  unsigned long mod = 0;
+  unsigned long long fraction = 0;
+  unsigned long long mod = 0;
   unsigned exp = 0;
   ieee_f f1_b,f2_b,f_b;
   f_b.f = 0;
@@ -140,8 +140,8 @@ float fdiv(float f1,float f2){
     return f_b.f;
   }
 
-  fraction = ((unsigned long)(0x800000+f1_b.b.fraction) << 25) / (unsigned long)(0x800000+f2_b.b.fraction);
-  mod = ((unsigned long)(0x800000+f1_b.b.fraction) << 25) % (unsigned long)(0x800000+f2_b.b.fraction);
+  fraction = ((unsigned long long)(0x800000+f1_b.b.fraction) << 25) / (unsigned long long)(0x800000+f2_b.b.fraction);
+  mod = ((unsigned long long)(0x800000+f1_b.b.fraction) << 25) % (unsigned long long)(0x800000+f2_b.b.fraction);
   exp = f1_b.b.exp - f2_b.b.exp + 127;
   if (f1_b.b.exp == 0) exp = 0;
